requirements.txt
altgraph==0.17.4
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.1.1
deep-translator==1.11.4
deepl==1.18.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.23.0
hyperframe==5.2.0
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
libretranslatepy==2.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.0.8
nbformat==5.10.4
numba==0.60.0
numpy~=1.26.4
packaging==24.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
pycparser==2.22
pydub==0.25.1
pygame==2.6.0
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.19.0
scikit-learn==1.5.1
scipy~=1.13.1
setuptools==70.2.0
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
srt==3.5.3
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
translate==3.6.1
typing_extensions==4.12.2
urllib3==2.2.2
vosk==0.3.45
Wave==0.0.2
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0
whisper==1.1.10
wit==6.0.1
torch~=2.3.1
transformers~=4.42.4
charset-normalizer
defusedxml~=0.7.1
pip~=24.1.2
pexpect~=4.2.1
regex~=2021.7.6
astunparse~=1.6.3
fsspec~=2024.6.1
aiohttp~=3.9.5
ujson~=4.0.2
pycares~=4.2.0
networkx~=3.2.1
pandas~=2.2.2
astroid~=1.5.2
zipp~=3.19.2
tokenizers~=0.19.1
safetensors~=0.4.3
huggingface-hub~=0.24.0
torchaudio~=2.3.1
accelerate~=0.32.1
psutil~=6.0.0
langid~=1.1.6
speechbrain~=1.0.0
langdetect~=1.0.9
HyperPyYAML~=1.2.2
sentencepiece~=0.2.0
streamlit~=1.37.1
rich~=13.7.1